Move ReactProps inside common
